Rename presto references to Trino, upgrade SPI
