--- src/main/resources/META-INF/services/io.prestosql.spi.Plugin
@@ -1,1 +1,0 @@
-com.github.presto.querylog.QueryLogPlugin
