+++ src/main/resources/META-INF/services/io.trino.spi.Plugin
@@ -1,0 +1,1 @@
+com.github.trino.querylog.QueryLogPlugin
